Move missing permissions warning to components
