+++ src/popup/components/missing-permissions-alert.tsx
@@ -1,8 +1,8 @@
 import browser from 'webextension-polyfill'
 import { useAsyncFn } from 'react-use'
-import { Button } from '../components/button'
+import { Button } from './button'
 import { useEffect } from 'react'
-import { Icon } from '../components/icon'
+import { Icon } from './icon'
 import AlertIcon from '@fortawesome/fontawesome-free/svgs/solid/triangle-exclamation.svg'
 import { css } from '@emotion/react'
 import { palette } from '../theme'
